hust-captcha-resolver/requirements.txt
anyio==4.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.1.1
cycler==0.12.1
einops==0.2.0
exceptiongroup==1.2.2
fastapi==0.85.0
filelock==3.16.1
fonttools==4.57.0
gdown==3.11.0
h11==0.14.0
idna==3.10
imageio==2.35.1
imgaug==0.4.0
importlib_resources==6.4.5
kiwisolver==1.4.7
lazy_loader==0.4
lmdb==1.0.0
matplotlib==3.7.5
networkx==3.1
numpy==1.24.4
opencv-python==4.11.0.86
packaging==24.2
Pillow==9.5.0
prefetch-generator==1.0.1
pydantic==1.10.21
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==0.20.0
python-multipart==0.0.5
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.32.3
scikit-image==0.21.0
scipy==1.10.1
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
starlette==0.20.4
tifffile==2023.7.10
torch==1.8.1
torchvision==0.9.1
tqdm==4.67.1
typing_extensions==4.13.1
urllib3==2.2.3
uvicorn==0.18.3
vietocr==0.3.2
zipp==3.20.2
